Update replace.md
Hello,
I kindly suggest to replace the current documentation of this function 'Replace' by this proposal. I published it on the Community Forum (see here : https://community.jeedom.com/t/tuto-de-la-bonne-utilisation-de-loutil-remplacer/125294/1), and all comments I received was positive.
It's seems obvious that the official documentation is currently not enough precise and clear, and could be confused. 
As this function is very powerful, but also dangerous for the health system due to is deeply commitment, I think it's very important to be accurate at all steps for everybody for a good use of it.
Thanks,
Daniel